requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
django==5.0.4
    # via
    #   -r requirements.in
    #   djangorestframework
django-rest-framework==0.1.0
    # via -r requirements.in
djangorestframework==3.15.1
    # via django-rest-framework
sqlparse==0.5.0
    # via django
typing-extensions==4.11.0
    # via asgiref
